refactor(Worklets): split Serializable classes into dedicated files
