+++ packages/react-native-worklets/Common/cpp/worklets/SharedItems/MemoryManager.cpp
@@ -1,7 +1,7 @@
 #include <jsi/jsi.h>
 #include <react/debug/react_native_assert.h>
 #include <worklets/SharedItems/MemoryManager.h>
-#include <worklets/SharedItems/Serializable.h>
+#include <worklets/SharedItems/Serializable/Serializable.h>
 #include <worklets/WorkletRuntime/WorkletRuntime.h>
 
 namespace worklets {
