+++ packages/react-native-worklets/Common/cpp/worklets/SharedItems/Serializable/SerializableRemoteFunction.h
@@ -1,9 +1,10 @@
 #pragma once
 
 #include <jsi/jsi.h>
-#include <worklets/SharedItems/Serializable.h>
+#include <worklets/SharedItems/Serializable/Serializable.h>
 #include <worklets/Tools/JSScheduler.h>
 #include <worklets/Tools/RNRuntimeStatus.h>
+#include <worklets/WorkletRuntime/RuntimeData.h>
 
 #include <memory>
 #include <string>
@@ -17,29 +18,29 @@ class SerializableRemoteFunction : public Serializable,
  public:
   /** Creates RN Runtime Remote Function. */
   SerializableRemoteFunction(
-      jsi::Runtime &rnRuntime,
+      facebook::jsi::Runtime &rnRuntime,
       const std::string &name,
-      jsi::Function &&function,
+      facebook::jsi::Function &&function,
       const std::shared_ptr<JSScheduler> &jsScheduler,
       const std::shared_ptr<RNRuntimeStatus> &rnRuntimeStatus)
       : Serializable(ValueType::RemoteFunctionType),
         hostRuntime_(&rnRuntime),
         hostRuntimeId_(RuntimeData::rnRuntimeId),
-        function_(std::make_unique<jsi::Value>(rnRuntime, std::move(function))),
+        function_(std::make_unique<facebook::jsi::Value>(rnRuntime, std::move(function))),
         name_(name),
         jsScheduler_(jsScheduler),
         rnRuntimeStatus_(rnRuntimeStatus) {}
 
   /** Creates Worklet Runtime Remote Function. */
   SerializableRemoteFunction(
-      jsi::Runtime &workletRuntime,
+      facebook::jsi::Runtime &workletRuntime,
       const std::string &name,
-      jsi::Function &&function,
+      facebook::jsi::Function &&function,
       RuntimeData::RuntimeId hostRuntimeId)
       : Serializable(ValueType::RemoteFunctionType),
         hostRuntime_(&workletRuntime),
         hostRuntimeId_(hostRuntimeId),
-        function_(std::make_unique<jsi::Value>(workletRuntime, std::move(function))),
+        function_(std::make_unique<facebook::jsi::Value>(workletRuntime, std::move(function))),
         name_(name),
         jsScheduler_(nullptr),
         rnRuntimeStatus_(nullptr) {}
@@ -53,7 +54,7 @@ class SerializableRemoteFunction : public Serializable,
       const std::shared_ptr<Serializable> &resolveValue,
       const std::shared_ptr<RuntimeManager> &runtimeManager);
 
-  jsi::Value toJSValue(jsi::Runtime &rt) override;
+  facebook::jsi::Value toJSValue(facebook::jsi::Runtime &rt) override;
 
   [[nodiscard]] bool isHostedOnRNRuntime() const noexcept {
     return hostRuntimeId_ == RuntimeData::rnRuntimeId;
@@ -64,12 +65,25 @@ class SerializableRemoteFunction : public Serializable,
   }
 
  private:
-  jsi::Runtime *hostRuntime_;
+  facebook::jsi::Runtime *hostRuntime_;
   const RuntimeData::RuntimeId hostRuntimeId_;
-  std::unique_ptr<jsi::Value> function_;
+  std::unique_ptr<facebook::jsi::Value> function_;
   const std::string name_;
   const std::shared_ptr<JSScheduler> jsScheduler_;
   const std::shared_ptr<RNRuntimeStatus> rnRuntimeStatus_;
 };
+
+facebook::jsi::Value makeRNRuntimeSerializableRemoteFunction(
+    facebook::jsi::Runtime &rnRuntime,
+    const std::string &name,
+    const facebook::jsi::Function &function,
+    const std::shared_ptr<JSScheduler> &jsScheduler,
+    const std::shared_ptr<RNRuntimeStatus> &rnRuntimeStatus);
+
+facebook::jsi::Value makeWorkletRuntimeSerializableRemoteFunction(
+    facebook::jsi::Runtime &workletRuntime,
+    const std::string &name,
+    const facebook::jsi::Function &function,
+    RuntimeData::RuntimeId hostRuntimeId);
 
 } // namespace worklets
