+++ packages/react-native-worklets/Common/cpp/worklets/SharedItems/Synchronizable.h
@@ -2,7 +2,7 @@
 
 #include <jsi/jsi.h>
 #include <react/debug/react_native_assert.h>
-#include <worklets/SharedItems/Serializable.h>
+#include <worklets/SharedItems/Serializable/Serializable.h>
 #include <worklets/SharedItems/SynchronizableAccess.h>
 
 #include <memory>
